refactor(nika-runtime): the crate-root tests get their own file — headroom before #291
lib.rs sat at 1461/1500 — and the #291 recover-await design lands
exactly in this file's settle spine. The 272-line test module moves to
src/tests.rs (mod tests; · super still resolves to the crate root —
semantics unchanged, raw-string fixtures byte-preserved). lib.rs:
1461 → 1125.

105 lib tests green · workspace suite green · public API proven
byte-identical · clippy -D clean.
