+++ crates/nika-runtime/src/tests.rs
@@ -1,0 +1,288 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2024-2026 SuperNovae Studio <contact@supernovae.studio>
+
+//! The crate-root test module — moved out of `lib.rs` 2026-07-10 at
+//! 1461/1500 LOC (the #291 recover-await design lands in that file's
+//! settle spine; the headroom must exist BEFORE the feature). `super`
+//! still resolves to the crate root — semantics unchanged.
+
+#![allow(clippy::expect_used, clippy::unwrap_used, clippy::panic)]
+
+use super::*;
+
+#[test]
+fn runtime_config_default_is_wave_width_seed_zero() {
+    let cfg = RuntimeConfig::default();
+    assert!(cfg.wave_parallelism.is_none());
+    assert_eq!(cfg.jitter_seed, 0);
+}
+
+#[test]
+fn envelope_values_carries_typed_defaults_and_containers() {
+    // The v1 string-only view dropped typed list defaults — the
+    // value model must carry them (for_each collections · spec 03).
+    let yaml = r#"
+nika: v1
+workflow: vals
+env:
+  API_BASE: "https://api.example.test"
+vars:
+  plain: "text"
+  urls: ["a", "b"]
+  topic: { type: string, default: "news" }
+tasks:
+  - id: t
+    exec: { command: "true" }
+"#;
+    let wf = nika_schema::parse(
+        yaml,
+        nika_schema::FileId::new(0),
+        nika_schema::ParseMode::Strict,
+    )
+    .expect("parses");
+    let (vars, env, name) = envelope_values(&wf, &BTreeMap::new());
+    assert_eq!(name, "vals");
+    assert_eq!(
+        env["API_BASE"],
+        Value::String("https://api.example.test".into())
+    );
+    assert_eq!(vars["plain"], Value::String("text".into()));
+    assert_eq!(vars["urls"], serde_json::json!(["a", "b"]));
+    assert_eq!(vars["topic"], Value::String("news".into()));
+}
+
+#[test]
+fn typed_output_type_mismatch_is_a_var009() {
+    // `outputs.n: { type: string }` — when the resolved value is a number
+    // the callable contract is broken (spec 01 §engine-MUST rule 6).
+    let yaml = r#"
+nika: v1
+workflow: typed-out
+tasks:
+  - id: t
+    invoke: { tool: "nika:jq", args: { input: { x: 42 }, expression: ".x" } }
+outputs:
+  n:
+    value: ${{ tasks.t.output }}
+    type: string
+"#;
+    let wf = nika_schema::parse(
+        yaml,
+        nika_schema::FileId::new(0),
+        nika_schema::ParseMode::Strict,
+    )
+    .expect("parses");
+    // A number where `string` is declared → NIKA-VAR-009.
+    let bad = BTreeMap::from([("n".to_owned(), serde_json::json!(42))]);
+    let v =
+        first_output_type_violation(&wf, &bad).expect("number vs declared string is a violation");
+    assert_eq!(v.name, "n");
+    assert_eq!(v.expected, "string");
+    assert_eq!(v.actual, "number");
+    // The declared type → no violation.
+    let good = BTreeMap::from([("n".to_owned(), serde_json::json!("hello"))]);
+    assert!(first_output_type_violation(&wf, &good).is_none());
+    // An unresolved output (omitted upstream) is NOT a type error.
+    assert!(first_output_type_violation(&wf, &BTreeMap::new()).is_none());
+}
+
+#[test]
+fn value_matches_vartype_lenient_floats_strict_cross_type() {
+    use serde_json::json;
+    // integer: whole floats OK, fractional rejected, numeric STRING rejected.
+    assert!(value_matches_vartype(&json!(42), VarType::Integer));
+    assert!(value_matches_vartype(&json!(42.0), VarType::Integer));
+    assert!(!value_matches_vartype(&json!(42.5), VarType::Integer));
+    // number: any JSON number, but NOT a numeric string.
+    assert!(value_matches_vartype(&json!(42), VarType::Number));
+    assert!(!value_matches_vartype(&json!("42"), VarType::Number));
+    // array vs object are distinct.
+    assert!(value_matches_vartype(&json!([1, 2]), VarType::Array));
+    assert!(!value_matches_vartype(&json!({}), VarType::Array));
+    assert!(value_matches_vartype(&json!({ "k": 1 }), VarType::Object));
+    assert!(value_matches_vartype(&json!("x"), VarType::String));
+    assert!(value_matches_vartype(&json!(true), VarType::Boolean));
+}
+
+/// A recovered success emits `task_recovered` BEFORE the terminal
+/// `task_completed` (engine#301 · D-2026-07-08-N4 sequence lock:
+/// `… > task_recovered > task_completed`) — and carries WHAT it
+/// recovered from as a `code` field. A clean success emits no such
+/// frame (pinned by every other Success test in this file).
+#[test]
+fn recovered_success_emits_task_recovered_before_completed() {
+    let ran = task::RanTask {
+        note: "exec · sh".to_owned(),
+        retries: Vec::new(),
+        agent_events: Vec::new(),
+        duration_ms: 0,
+        result: task::RunResult::Success {
+            value: Value::Number(99.into()),
+            tokens: None,
+            recovered_from: Some("NIKA-EXEC-001".to_owned()),
+            warning: None,
+            cost_usd: None,
+            cost_unpriced: None,
+        },
+    };
+    let mut ok = true;
+    let mut stamper = DeterministicStamper::new();
+    let mut sink = VecSink::new();
+    settle_ran("risky", ran, None, &mut ok, &mut stamper, &mut sink);
+
+    let kinds: Vec<EventKind> = sink.events().iter().map(|e| e.kind).collect();
+    let rec = kinds
+        .iter()
+        .position(|k| *k == EventKind::TaskRecovered)
+        .expect("a TaskRecovered frame");
+    let done = kinds
+        .iter()
+        .position(|k| *k == EventKind::TaskCompleted)
+        .expect("a TaskCompleted frame — completed STAYS the one success terminal");
+    assert!(rec < done, "task_recovered inserts BEFORE the terminal");
+
+    let frame = &sink.events()[rec];
+    assert!(
+        frame.fields.iter().any(|f| f.key == "code"
+            && matches!(&f.value, FieldValue::String(s) if s == "NIKA-EXEC-001")),
+        "the frame names what was recovered FROM"
+    );
+    assert!(ok, "a recovered task is a SUCCESS at workflow level");
+}
+
+/// A settled success carrying an OBS-E `warning` puts it on the
+/// `TaskCompleted` frame as a `warning` field — the wiring proof that
+/// the dispatch's diagnostic actually reaches the event stream.
+#[test]
+fn obs_e_warning_rides_task_completed() {
+    let ran = task::RanTask {
+        note: "infer · gemini/flash".to_owned(),
+        retries: Vec::new(),
+        agent_events: Vec::new(),
+        duration_ms: 0,
+        result: task::RunResult::Success {
+            value: Value::String(String::new()),
+            tokens: Some(84),
+            recovered_from: None,
+            warning: Some("infer produced an empty answer · …".to_owned()),
+            cost_usd: Some(0.0125),
+            cost_unpriced: None,
+        },
+    };
+    let mut ok = true;
+    let mut stamper = DeterministicStamper::new();
+    let mut sink = VecSink::new();
+    settle_ran("think", ran, None, &mut ok, &mut stamper, &mut sink);
+
+    let completed = sink
+        .events()
+        .iter()
+        .find(|e| e.kind == EventKind::TaskCompleted)
+        .expect("a TaskCompleted frame");
+    let warning = completed
+        .fields
+        .iter()
+        .find(|f| f.key == "warning")
+        .expect("the warning field rides the success frame");
+    assert!(
+        matches!(&warning.value, FieldValue::String(s) if s.contains("empty answer")),
+        "the diagnostic text is carried verbatim"
+    );
+    // Real spend rides the same frame · absent-when-unpriced is pinned
+    // by the sibling test below (its cost_usd is None · no field).
+    let cost = completed
+        .fields
+        .iter()
+        .find(|f| f.key == "cost_usd")
+        .expect("the cost_usd field rides the priced success frame");
+    assert!(
+        matches!(&cost.value, FieldValue::Float(c) if (*c - 0.0125).abs() < f64::EPSILON),
+        "the priced spend is carried verbatim"
+    );
+}
+
+/// The common path · a success with no OBS-E diagnostic emits NO
+/// `warning` field (zero false-alarm noise on the happy stream).
+#[test]
+fn no_warning_field_on_a_clean_success() {
+    let ran = task::RanTask {
+        note: "exec · true".to_owned(),
+        retries: Vec::new(),
+        agent_events: Vec::new(),
+        duration_ms: 0,
+        result: task::RunResult::Success {
+            value: Value::String("ok".to_owned()),
+            tokens: None,
+            recovered_from: None,
+            warning: None,
+            cost_usd: None,
+            cost_unpriced: None,
+        },
+    };
+    let mut ok = true;
+    let mut stamper = DeterministicStamper::new();
+    let mut sink = VecSink::new();
+    settle_ran("t", ran, None, &mut ok, &mut stamper, &mut sink);
+
+    let completed = sink
+        .events()
+        .iter()
+        .find(|e| e.kind == EventKind::TaskCompleted)
+        .expect("a TaskCompleted frame");
+    assert!(
+        !completed.fields.iter().any(|f| f.key == "warning"),
+        "no warning on a clean success"
+    );
+    assert!(
+        !completed.fields.iter().any(|f| f.key == "cost_usd"),
+        "an unpriced success carries NO cost field — absent is honest, never a fake zero"
+    );
+    assert!(
+        !completed.fields.iter().any(|f| f.key == "cost_unpriced"),
+        "an exec success is not COST-unpriced — no reason noise on verbs that spend nothing"
+    );
+}
+
+/// The WHY channel: an unpriced INFER success carries the reason —
+/// `unknown` is never masked (a local model says « local compute ·
+/// not priced », never a blank).
+#[test]
+fn cost_unpriced_reason_rides_task_completed() {
+    let ran = task::RanTask {
+        note: "infer · ollama/llama3.2".to_owned(),
+        retries: Vec::new(),
+        agent_events: Vec::new(),
+        duration_ms: 0,
+        result: task::RunResult::Success {
+            value: Value::String("bonjour".to_owned()),
+            tokens: Some(12),
+            recovered_from: None,
+            warning: None,
+            cost_usd: None,
+            cost_unpriced: Some(nika_types::cost::UnpricedReason::LocalModel),
+        },
+    };
+    let mut ok = true;
+    let mut stamper = DeterministicStamper::new();
+    let mut sink = VecSink::new();
+    settle_ran("ask", ran, None, &mut ok, &mut stamper, &mut sink);
+
+    let completed = sink
+        .events()
+        .iter()
+        .find(|e| e.kind == EventKind::TaskCompleted)
+        .expect("a TaskCompleted frame");
+    assert!(
+        !completed.fields.iter().any(|f| f.key == "cost_usd"),
+        "no fake zero next to the reason"
+    );
+    let reason = completed
+        .fields
+        .iter()
+        .find(|f| f.key == "cost_unpriced")
+        .expect("the WHY rides the frame");
+    assert!(
+        matches!(&reason.value, FieldValue::String(s) if s == "local_model"),
+        "snake_case wire form"
+    );
+}
